@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -60,7 +60,12 @@ dependencies:
   chewie: ^1.3.1
   path_provider: ^2.0.9
   open_file: ^3.2.1
-
+  font_awesome_flutter: ^10.1.0
+  percent_indicator: ^4.0.0
+  firebase_messaging: ^11.2.14
+  flutter_local_notifications: ^9.4.0
+  firebase_core: ^1.15.0
+  http: ^0.13.4
 
 
   # The following adds the Cupertino Icons font to your application.
