@@ -1,10 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:flutter_chat/ringy/domain/entities/chat_message/chat_message.dart';
+import 'package:flutter_chat/ringy/domain/entities/chat/chat_message.dart';
 import 'package:flutter_chat/ringy/presentation/core/widgets/encryption_utils.dart';
-import 'package:flutter_chat/ringy/resources/colors.dart';
-import 'package:flutter_chat/ringy/resources/constants.dart';
 import 'package:flutter_chat/ringy/resources/shared_preference.dart';
 import 'package:flutter_chat/ringy/resources/strings_en.dart';
 
@@ -24,19 +22,6 @@ class NormalMessageView extends StatelessWidget {
         messagesList[index].senderId?.id == Prefs.getString(Prefs.myUserId);
     var isMessageDeleted = messagesList[index].isDeleted == 1;
     return
-      // Row(
-      // mainAxisSize: MainAxisSize.min,
-      // // verticalDirection: VerticalDirection.down,
-      // children: [
-      //      if(messagesList[index].isInProgress!)   SizedBox(
-      //     width: 18,
-      //     height: 18,
-      //     child: CircularProgressIndicator(
-      //       color: isMyMessage ? RingyColors.lightWhite:RingyColors.primaryColor,
-      //       strokeWidth: 2,
-      //     ),
-      //   ),
-      //   if(messagesList[index].isInProgress!) const SizedBox(width: 8,),
         Flexible(
           child: Text(
             isMessageDeleted ? StringsEn.deleted: EncryptData.decryptAES(messagesList[index].message.toString(),
