@@ -1,17 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_chat/ringy/domain/entities/chat_message/chatusers/users_model.dart';
-import 'package:flutter_chat/ringy/domain/entities/chat_message/user_model.dart';
-import 'package:flutter_chat/ringy/presentation/core/utils/data_travel_model.dart';
-import 'package:flutter_chat/ringy/presentation/core/utils/helper_class.dart';
-import 'package:flutter_chat/ringy/presentation/core/widgets/encryption_utils.dart';
-import 'package:flutter_chat/ringy/presentation/core/widgets/image_or_first_character.dart';
+import 'package:flutter_chat/ringy/domain/entities/users/chatusers/users_model.dart';
 import 'package:flutter_chat/ringy/presentation/core/widgets/image_or_first_character_users.dart';
 import 'package:flutter_chat/ringy/resources/colors.dart';
-import 'package:flutter_chat/ringy/resources/constants.dart';
-import 'package:flutter_chat/ringy/resources/strings_en.dart';
-
-import '../../../../../routes/router.dart';
 
 class UserItemForwardTile extends StatelessWidget {
   final UsersList model;
@@ -51,16 +42,6 @@ class UserItemForwardTile extends StatelessWidget {
   }
 
   _nextPage(BuildContext context, UsersList model) {
-    // TmpDataTravel tmpDataTravel = TmpDataTravel();
-    // tmpDataTravel.name = model.name!;
-    // tmpDataTravel.image = model.userImage!;
-    // tmpDataTravel.recieverId = model.sId!;
-    // tmpDataTravel.isOnlineHide = "0";
-    // tmpDataTravel.isOnline = model.onlineStatus!;
-    // tmpDataTravel.mainUserId = model.userId!;
-    //
-    // context.pushRoute(ChatScreenRoute(dataTravel: tmpDataTravel));
     context.popRoute(model);
-    // Navigator.pop(context, model);
   }
 }
