@@ -1,5 +1,7 @@
 
 import 'package:flutter/material.dart';
+import 'package:flutter_chat/ringy/presentation/home/chat/group/groupsList/groups_list_page.dart';
+import 'package:flutter_chat/ringy/presentation/home/settings/profile.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:flutter_chat/ringy/presentation/core/widgets/error_retry_widget.dart';
 import 'package:flutter_chat/ringy/presentation/home/chat/o2o/o2o_users/widgets/o2o_users.dart';
@@ -7,31 +9,30 @@ import 'package:flutter_chat/ringy/resources/colors.dart';
 import 'package:flutter_chat/ringy/resources/strings_en.dart';
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key? key}) : super(key: key);
+  int _currentIndex;
+
+
+  HomePage(this._currentIndex, {Key? key}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
 }
 
 class _HomePageState extends State<HomePage> {
-  int _currentIndex = 0;
+
 
   _getDrawerItemWidget(int pos) {
     switch (pos) {
       case 0:
         return const O2OUsersPage(false);
       case 1:
-        return const Center(
-          child: Text('Group Users'),
-        );
+        return const GroupsListPage(false);
       case 2:
         return const Center(
           child: Text('Calls'),
         );
       case 3:
-        return const Center(
-          child: Text('Rings'),
-        );
+        return  Profile();
       default:
         return ErrorRetryWidget("error!", () => {
         });
@@ -43,9 +44,9 @@ class _HomePageState extends State<HomePage> {
     return Scaffold(
       backgroundColor: Colors.blueGrey[100],
       // body: screens[_currentIndex],
-      body: _getDrawerItemWidget(_currentIndex),
+      body: _getDrawerItemWidget(widget._currentIndex),
       bottomNavigationBar: BottomNavigationBar(
-        currentIndex: _currentIndex,
+        currentIndex: widget._currentIndex,
         items: <BottomNavigationBarItem>[
           BottomNavigationBarItem(
               activeIcon: SvgPicture.asset(
@@ -77,7 +78,7 @@ class _HomePageState extends State<HomePage> {
         selectedFontSize: 12,
         unselectedFontSize: 12,
         unselectedItemColor: Colors.grey,
-        onTap: (index) => setState(() => _currentIndex = index),
+        onTap: (index) => setState(() => widget._currentIndex = index),
       ),
     );
   }
