@@ -5,25 +5,21 @@ class Constants {
 
 
 
-  static const int NORMAL_MSG = 0;
-  static const int IMAGE_MSG = 1;
-  static const int FILE_MSG = 2;
-  static const int CALL_MSG = 3;
-  static const int LINK_MSG = 4;
-  static const int VIDEO_MSG = 5;
-  static const int AUDIO_MSG = 6;
-  static const int LOCATION_MSG = 7;
-  static const int STICKER_MSG = 8;
+  static const int normalMSG = 0;
+  static const int imageMSG = 1;
+  static const int fileMSG = 2;
+  static const int callMSG = 3;
+  static const int linkMSG = 4;
+  static const int videoMSG = 5;
+  static const int audioMSG = 6;
+  static const int locationMSG = 7;
+  static const int stickerMSG = 8;
 
 
   static const int passwordLength = 6;
 
-
-
-
   static const String projectId = "5d4c07fb030f5d0600bf5c07";
-
-  static const String SOCKET_URL = "https://peekvideochat.com:22000";
+  static const String socketURL = "https://peekvideochat.com:22000";
 
   static const int connectPage = 0;
   static const int friendsPage = 1;
